Added test to run app and check for exception
